Migrate submit transaction endpoint to new action pattern
