+++ services/horizon/internal/actions.go
@@ -39,9 +39,3 @@ func (action TradeIndexAction) Handle(w http.ResponseWriter, r *http.Request) {
 	ap.Prepare(w, r)
 	ap.Execute(&action)
 }
-
-func (action TransactionCreateAction) Handle(w http.ResponseWriter, r *http.Request) {
-	ap := &action.Action
-	ap.Prepare(w, r)
-	ap.Execute(&action)
-}
